--- packages/perseus/src/widget-ai-utils/iframe/prompt-utils.test.ts
@@ -1,13 +1,0 @@
-import {getPromptJSON} from "./prompt-utils";
-
-describe("Iframe getPromptJSON", () => {
-    it("it returns JSON with the expected format and fields", () => {
-        const resultJSON = getPromptJSON();
-
-        expect(resultJSON).toEqual({
-            type: "iframe",
-            isSupported: false,
-            message: "",
-        });
-    });
-});
